Updates to reduce runtime
Made several tweaks in an attempt to reduce the time required to run the code. In the value_listfinder function, the rows that are combined to be added to the JSON are deleted from the EIA numpy array so that subsequent searches of the array are faster. This change doesn't seem to have improved runtimes with the small test JSON file, but with the full JSON database to populate, it might help
speed up total runtime.

Additionally, the call to the value_listfinder_filterformat function in value_listfinder was moved to outside the for loop, reducing the number of times the _filterformat function was called.


Former-commit-id: b81063a7cc41a4ff24fdf17c11026ffdeb9c837e