Removed fuel pump microsegments
Consistent with modifications made to the code that populates the JSON (mseg.py), the JSON file itself no longer has microsegments reflecting the fuel pump data in the EIA data file. These data are of a fundamentally different character compared to the main microsegment data, and since there are no existing measures that affect fuel pumps, they are excluded here.


Former-commit-id: 74ea97740a02b29f822641fe26af73b2abcb059b