Changed 'mseg.py' to accommodate 2nd heat changes
The previous commit updated the microsegments json to show “demand” and
“supply” technology sub-categories for secondary heating (as had
previously been shown only for the primary heating and cooling end
uses).  This commit updates the mseg.py file to accommodate these new
secondary heating sub-categories.

The main updates are to the nested for loop in the
“value_replacer_main” function, and to the “demand” and “supply” if
statements in the “json_translator” function.  In both cases, the
updates are needed to accommodate the special microsegment structure
for secondary heating in the “other fuel” type, which goes down one
sub-level further than other microsegments in the json.  This is
because secondary heating within the “other fuel” category is handled
on the technology sub-level and not the end use sub-level as for the
electricity, natural gas, and distillate fuel categories.

For the electricity, natural gas, and distillate fuel categories, the
new “demand” and “supply” subcategories for secondary heating end use
are handled just as they are for the primary heating and cooling end
uses.


Former-commit-id: 497e523f09e4988e130bba80fee2b731efd48a33