Revised parsing of EIA text file
In the future, EIA might restructure the data files used by mseg.py to define the supply microsegments. Because the columns with equipment stock and consumption data were originally indicated with numeric column positions, changes in the columns might go undetected and result in erroneous program outputs that aren't readily apparent. To improve code readability and induce an exception if there are changes in the heading names (but not positions), the numeric indices were replaced with column name strings.


Former-commit-id: cf9baabfc9423cb1bb5fd8f1388f77d328cfc9b2