Expanded cases in JSONTranslatorTest
To try and capture the new supply and demand microsegments that were sometimes causing errors in mseg.py, additional cases were added to the list 'ok_filters'. If the function json_translator is working as intended, all of these filters should match (pass assertEqual) and the json_translator function should not raise any exceptions when running the tests.


Former-commit-id: e7ced3716374a84bbff0564601a1025e1d8ccaad