Updated "json_translator" and associated tests
The “json_translator” function has been updated to be more compact and
flexible.

Given a list of json->txt microsegment name translation dicts and a
list of json names for a particular microsegment, the function runs
through the list of translation dicts and determines whether any of the
json names correspond to dict keys.  If so, the function adds the
values for the matched keys to an output list of filtering information
for later use in parsing the EIA .txt file.

Note that the output list has two elements, the first of which includes
filtering information for “supply” microsegments, and the second of
which includes filtering information for “demand” microsegments.

The test for the updated “json_translator” function has been slightly
updated to reflect its modified output, which had previously consisted
of a tuple of filtering information for “demand” microsegments and a
list of filtering information for “supply” microsegments.  The
“filter_formatter” function and its test have also been updated to
reflect the modified “json_translator” output, which serves as its
input.


Former-commit-id: 093dd552e35776745f6c912439ccbf4c5a657c38