Write mseg.py output & minor fixes
The mseg.py routine now writes its updated microsegments dict to a
‘microsegments_out.json’ file, formatted with indent = 4 spaces for
easy readability.  This replaces the earlier “return” of the updated
dict.

In a test-write of the “microsegments_out.json” file, it was found that
deletion of matched numpy rows in “txt_parser” was causing inaccuracies
in updating “demand” microsegments.  Specifically, each of the “demand”
microsegments for a given census division, building type, fuel, and end
use must access the same “group_energy_base” information to apply the
thermal load component to - i.e., “all gas heating energy use in single
family homes in New England”.  However, this base information was being
deleted from the numpy array after the first “demand” microsegment was
updated, leaving zeros for all remaining “demand” microsegments in that
census division, building type, fuel type, and end use combination.  To
fix this, the “txt_parser” no longer deletes rows when a “demand” numpy
array is being parsed in a “Record” operation (signified by an
“EIA_Demand” file_type input and “Record” command_string input).

Finally, updated microsegment stock and energy values are now rounded
to 4 decimal places in the “merge_sum” function, to prevent very long
numbers from being written to the “microsegments_out.json” file.  As
this rounding used to be done in “mseg_test.py,” that part of the test
code is no longer needed and has been removed.


Former-commit-id: 4eb65d5f883f078ac49629d19a0e3ee8fd6d1be7