Removed unneeded prototype tests
Removed two tests, one to explore the structure of the input JSON file to ensure that it has the anticipated hierarchy/structure, and one to test for the type (byte or string) of input string data. The former was removed because mseg.py is designed to traverse the JSON and assign the data appropriately regardless of the order in which the microsegment categories appear, which is a more flexible approach and also obviates the need for a test of the JSON database structure. The latter was removed because the handling of text by numpy.genfromtxt with dtype=None is unlikely to move away from bytecode strings.


Former-commit-id: 31eb7347a6a386dff2f8b202b2aa022cbb7932cb