Updated stock_consume_select function
The stock_consume_select function was updated to simplify its structure with regard to handling of supply and demand data. Because demand microsegment data might need to be reused later for other thermal load components (e.g. heating-specific microsegments will apply to losses via conduction through windows and [opaque] envelope), the function must either not eliminate rows used when processing demand data or return the data array only when it is updated. The former approach is now used, where the rows_to_remove list is only added to when the 'EIA_Supply' case text is passed to the function.

In addition to the revisions to the function itself, a test was added to demonstrate that when passing file_type values other than 'EIA_Supply', the stock_consume_select function returns a data array identical to that originally sent to the function.


Former-commit-id: 497b3516a52ea0c2a112dadff2181cf805ff0061