Initial run.py scheme
This is an initial attempt to sketch out measures in an Object Oriented
Modeling (OOM) scheme.  Measure and microsegment information is
imported from JSON files.  Measures objects are instantiated under the
“Measure” class, and methods on the measures include finding
corresponding microsegments and partitioning them into “competed” and
“non-competed” sub-segments.  Note that for each projection year, a
measure will only have access to the “competed” sub-segment unless a
“technical potential” adoption rule is specified, in which case the
entire microsegment is accessed.

The “Engine” class is then used to simulate the adoption of active
measures, given the user-defined adoption rule.


Former-commit-id: a1da1a199b7d83ba2953f4ca1ffa52d806e45bd7