Added measure attributes, "calc_metrics" method, and "main" function
This commit adds the initialization of several measure object
attributes that must be defined in the “run.py” routine on top of the
attributes already imported from “measures.json”.

Additionally, it adds a third method to the “Measure” class
(“calc_metrics”) and fleshes out how the attributes of active measure
objects are to be updated as part of the “adopt_active” method in the
“Engine” class.

Finally, a “main” function is added that executes the user-defined
measure adoption scheme.  Note that both the adoption scheme and a
decision rule scheme are now inputs to the “adopt_active” method.  The
latter will only be needed for more complex adoption scenarios.


Former-commit-id: caa58d333832fe571238989dabb67e95b9b0d661