Remove 'years' key from JSON database
Since the years associated with the data are now included in stock and consumption dicts at the leaf/terminal nodes in the microsegments JSON database, the 'years' key to be filled at the first level of the JSON database is no longer needed. The 'years' and companion 'data' keys were removed from the database, eliminating that level in the hierarchy. The corresponding lines of code in mseg.py to populate the 'years' portion of the database and support the 'data' key were also removed.


Former-commit-id: 73b8d85153adedd27d2b2d0c345857f78ea7619f