Updated measure data formatting
Changed the order of performance levels for the mini-split heat pump measure so that climate zone comes first, followed by end use (heating or cooling). This order follows the structure in the microsegments JSON database, and should simplify the process of comparing microsegment energy use and stock data to the performance cited in the measures.


Former-commit-id: 1aa10cbf688b9464460061c9f94a2437bcc1e9e6