Climate conversions & test reflects newest mseg JSON format
The “clim_converter” and supporting “merge_sum” functions in “mseg.py”
were still expecting input stock/energy values in list format and
outputting climate-converted stock/energy values in this same format.
With this commit, the “clim_converter” function now accepts and outputs
stock and energy information in a dict format with projection years as
the keys; this is consistent with how stock and energy projection
information is organized in the current “microsegments.json” file.


Former-commit-id: 78fa2b6401213c8def592b79fa1ae3d870e60261