Adjusted handling of relative performance calculation
In most cases, a measure will have positive relative performance over a
baseline if its absolute performance value is higher (i.e., COP 4 vs.
COP 3).

However, this is not always the case.  For example, a measure might
reduce baseline infiltration for a home from 13 ACH50 to 1 ACH50, and
this should be identified as a positive relative performance for the
measure.

This commit uses a measure’s performance units to determine whether the
measure’s relative performance calculation must be inverted to yield
the correct result.  These special inversion cases are defined at the
beginning of “run.py” in an “inverted_relperf_list”, which may be added
to over time.

In addition, the “1 - ” has been removed from the relative performance
calculation and subsequent “mseg_efficient” calculation in
“partition_microsegment” because it was uneccessary when used in both
these calculations.

Tests in “run_test.py” have been updated in accordance with the above
changes.


Former-commit-id: d7dd3eaea4f28540956b3d9c4e3c6ba033dd437f