Record all contributing microsegment info. for use in measure staging
All microsegments that contribute to a measure’s master microsegment
are now recorded in “mseg_find_partition” for later use in determining
overlaps between microsegments of multiple competing measures.

Stock, energy, and cost data for the contributing microsegments are
recorded along with the key chain that is used to access this
information from an input dictionary; each key chain becomes a unique
identifier for the microsegment that is used to determine its overlap
with microsegments for other measures.

Microsegment overlaps figure into the measure staging process, whereby
measures that are adopted first capture a certain portion of the energy
market for measures with overlapping markets that are subsequently
adopted.  As an example, if it is determined residential LEDs will be
adopted before CFLs and the two share the same residential lighting
energy market, this market must be reduced by the amount of the LED
lighting savings before subsequently calculating CFL savings.

With this update, the “mseg_find_partition’’ function now returns a
list, where the first list element contains a measure’s master
microsegment information and the second list element contains all
contributing microsegment information.  The
FindPartitionMicrosegmentTest in “run_test.py” has been adjusted
accordingly.


Former-commit-id: 96b3c6ee9ef51bb37dc8a8d81d01122ec7d7ea0b