Add info. on energy/carbon costs, site-source conversions, and carbon intensities
This commit adds a .txt file to the repository with AEO-based
projections of energy costs, site to source conversion factors, and
carbon dioxide emission intensities for electric, natural gas, and
“other” fuels.  The .txt file also includes projections of carbon costs
drawn from the U.S. Social Cost of Carbon documentation*.

This information will be folded into the construction of measure master
microsegment data in the “run.py” routine.  Currently, the .txt file is
constructed and updated via an associated Excel file; in the future, it
might be updated with a python routine.

* Using 3% discount rate, interpolating for years between those
projected on page 3 of this document:
https://www.whitehouse.gov/sites/default/files/omb/inforeg/social_cost_o
f_carbon_for_ria_2013_update.pdf

Former-commit-id: b2fb47b3cbdf39e052adbfe544a4ff3a8d137f9d