Calculate measure prioritization metrics
The “calc_metric_update” function in “run.py” now yields a measure’s
internal rate of return (IRR), simple payback, cost of conserved energy
(CCE), and cost of conserved carbon (CCC).

IRR and simple payback are each calculated under two scenarios: 1)
considering capital and energy cost cash flows, and 2) considering
capital, energy, and carbon cost cash flows.  CCE is calculated with
and without additional benefits assumed from carbon cost savings, and
CCC is likewise calculated with and without additional benefits assumed
from energy cost savings.  Where appropriate, a function input for
discount rate is used in the above calculations.

Metric outputs depend on a measure’s stock, energy/energy cost, and
carbon/carbon cost savings, which are drawn from the measure’s master
microsegment information (established by the “mseg_find_partition”
function).  These underlying savings figures are also included in the
output of the “calc_metric_update” function.

Note that the formats of the stock, energy, and cost variables in
“calc_metric_update” will not always be the same.  For example, if the
user places a distribution on the capital costs for a measure but not
on the performance, capital costs will be in list format while annual
energy/carbon savings will be point values.  To simplify subsequent
calculations that combine these variables, each is specified as a list
of length N, where N is the maximum list length across all the
variables*.

All lists used in the “mseg_find_partition” and “calc_metric_update”
functions have been reformatted as numpy arrays, which are as easily
combined with point values as they are with other arrays (i.e., we
might have var1 = [1,2,3] and var2 = 3 and want var1 * var2 = [3,6,9]).
 This simplifies the coding of operations between variables that may be
all arrays; all point values; or a mix of both.

New tests have been added to “run_test.py” that reflect the above
changes to “calc_metric_update” as well as the addition of supporting
functions “metric_update” and “payback”.  It is noted that the test for
“calc_metric_update” (“PrioritizationMetricsTest”) is incomplete, as it
does not consider a case where measure master microsegment information
includes lists of values.  This case will be added to the test in a
future commit.

* For example, if the initial capital investment ($) for a measure is
array([10, 15, 20]) and the annual energy savings (mmBTU) is 500, the
annual energy savings will be redefined as array([500, 500, 500]), to
match the formatting and dimensions of the capital investment variable.


Former-commit-id: 7bc27c122746a45825c00363dd3668089894ace3