Added test for the two "fill_years_" functions in "mseg_techdata.py"
This test covers both the “fill_years_nlt” and “fill_years_lt”
functions in “mseg_techdata.py”.  The “fill_years_nlt” function
reconstructs EIA performance and cost projections for non-lighting
technologies into a list of performance and cost output dicts; these
dicts include information for each projection year used in the
“mseg.py” routine. The “fill_years_nlt” does the same for lighting
technologies, also adding in lifetime information (which the EIA does
not break out by year for the non-lighting technology cases).

For the “fill_years_nlt” function, there are special cases for
refrigerator and freezer technologies where the input EIA information
to reconstruct will include multiple rows for a given projection year,
due to multiple configurations for these technologies (i.e., compact
and upright freezers).  In these cases, the data for the multiple
configurations are averaged into a single configuration for use in the
“fill_years_nlt” and supporting “stitch” routines.  This capability is
tested for the freezers case.

Another part of the test checks that when the input array of
information includes multiple rows with the same projection year and a
special refrigerators/freezers technology case is not present, the
“fill_years_nlt” function yields an error.  The “fill_years_lt” is
tested in a similar manner.

A few minor adjustments have been made to the “mseg_techdata.py”
routine itself to ensure that this new test passes and to clarify some
comments.


Former-commit-id: 65cbb82a5b3ae1ee3f311ddbe08d9200b4a610f0