Added new cdiv-clim conversion .txt files
Two .txt files have been added that show the distribution of
residential and commercial square footage in each AIA climate zone
across the nine census divisions.  In the .txt arrays, climate zones
are represented in the columns and census divisions are represented in
the rows.  Each column’s values add to 1 (representing 100% of the
square feet in the climate zone).

These files are used to translate technology performance, cost, and
lifetime information from a census division breakdown to a climate zone
breakdown.


Former-commit-id: 8694ca5eddb36d2c7b29bc43e75cc20f2c0c337e