Update to fix ‘other’ fuel data
The regular expressions created by the filter_formatter function, which are used to select the appropriate data, will greedily match text following each entry. This meant that in cases where ‘other’ is selected as the fuel type, the regex would match text of a row of data for any of the 7 fuel types in the fueltype dict at any point beyond the number provided for the building type. The inclusion of ‘NG’ in the other fuel group meant that the regex would match natural gas technologies that had a technology type description that included ‘NG.’ Since ‘NG’ is only used as the fuel type for fuel pumps, it was removed from the ‘other fuel’ list. Similarly, the energy input to solar water heating systems from solar insolation is not relevant to current analyses (and if it is of interest later, it should be included as a separate fuel type) thus the ‘SL’ fuel type code was also removed from the ‘other fuel’ list.

These changes fix the incorrect data recorded for ‘other’ cooking and water heating in microsegments_out.json.


Former-commit-id: d89a9a5a6b92b7a13fad151e53cfcdd8ab53565f