Allow definition of secondary microsegments for a measure
To accommodate measures with end use interactions (e.g. improved
lighting efficiency affecting heating/cooling load), the microsegments
that contribute to a measure’s master microsegment are now defined as
either ‘primary’ or ‘secondary’ microsegments.

As an example, consider a measure that improves a building’s lighting
equipment efficiency.  The improvement reduces the amount of lighting
energy used by the building, but also reduces the amount of waste heat
the building’s lighting generates, thus increasing the building’s
heating load and decreasing the cooling load.  Here, the affected
lighting end use microsegments constitute a primary case, while the
heating and cooling load microsegments constitute a secondary case.
Note that for the secondary case, energy use, associated carbon
emissions and energy/carbon costs are affected without the addition of
any new heating/cooling equipment; thus, secondary microsegments do not
affect stock, stock cost, or stock lifetime calculations for a measure.

Secondary microsegments are constructed via dictionaries for each
relevant field in the measures JSON (e.g., ‘end_use,’ ‘fuel_type,’
‘technology_type,’ and ‘technology’).  Each dictionary includes a key
for ‘primary’ and ‘secondary’ microsegment information.  Energy savings
associated with each secondary microsegment are specified just as they
are for primary microsegments using the ‘energy_efficiency’ and
‘energy_efficiency_units’ fields in the measures JSON (also now
dictionaries).  A distribution may be placed on the energy efficiency
value for a given secondary microsegment, and values may be specified
separately for each secondary microsegment that a measure applies to if
needed.

A new sub-routine called ‘create_keychain’ was added to the
‘mseg_find_partition’ function in ‘run.py’ to accommodate the above
changes.  Given a set of primary or secondary microsegment information
as an input, the routine builds a list of relevant key chains to use in
parsing through the microsegments JSON and basecostperflife JSON data
(e.g., cdiv->bldg. type->fuel type->end use->tech. type).  Each key
chain is tagged as ‘primary’ or ‘secondary’ as appropriate.  The
function yields a combined list of primary and secondary key chains as
an output; these key chains are utilized by the remainder of the
‘mseg_find_partition’ routine much as they were before this commit.*

Corresponding changes to ‘run_test.py’ include a revised set of sample
measures in ‘FindPartitionMasterMicrosegmentTest’ that reflect the new
measure field structure; and a new test of the ‘create_keychain’
function.  Note that in the revised
‘FindPartitionMasterMicrosegmentTest,’ sample measure 3 in the
‘ok_measures’ list tests the case where lighting efficiency
improvements have an impact on heating/cooling load.

While this commit accommodates the direct specification of secondary
end use impacts in a measure’s definition, the code does not yet handle
cases where a secondary effect is not specified in the measure
definition but should still be represented.  A future commit will have
‘run.py’ flag such cases for commercial lighting measures and make
appropriate microsegment adjustments to ensure that the secondary
effect is indeed represented.

* The only changes to the rest of ‘mseg_find_partition’ are mainly
geared towards ensuring that stock, stock cost, and lifetime
information is not updated for secondary microsegments, which do not
affect these variables.


Former-commit-id: ed1cc422920d4a1d2f3207b8645d91f7b9e6fe62