Automatically handle secondary effects of lighting efficiency
The “mseg_find_partition” routine in “run.py” has been adjusted to
handle the case where a commercial lighting measure definition does not
include information about the secondary end use impacts of lighting
efficiency improvements (e.g., the impacts on heating and cooling
energy use).

In this case, the routine automatically adds the relevant secondary
effect using the thermal load components data that are available for
commercial buildings.  These load component data include the portion of
the commercial heating and cooling loads attributable to waste heat
from lights.  In the case of heating, the load is automatically scaled
up by the relative improvement in lighting efficiency; in the case of
cooling, the load is automatically scaled down by the same amount.

To test this case, an additional sample measure (“sample measure 8”)
has been added to the “FindPartitionMasterMicrosegmentTest” in
“run_test.py”.  Supporting additions to sample input microsegment
information for these tests were also required.

Note that automatic consideration of secondary end use impacts is not
possible for a residential lighting efficiency measure, as the thermal
load component attributable to waste heat from residential lights is
not currently known (and would likely be small anyway for single family
dwellings).


Former-commit-id: 01d5991cce49437f70c55f70ba93b95d564357a6