Establish framework for microsegment partitions
This commit mainly concerns the partition_microsegment function in
‘run.py’.  The function is reworked to lay out all possible ways that
an input microsegment may be further partitioned given the definition
of the measure that accesses it; these partitions are then used to
update the total size of the microsegment and the portion of it that
the measure competes for in a given year.

In total, there are 7 possible ways an input microsegment may be
partitioned in a given year of the modeling time horizon:

•	Portion of total stock that is new buildings

o	Portion of new buildings adopting baseline technology (#1)
o	Portion of new buildings adopting efficient technology (#2)

•	Portion of total stock that is existing buildings

o	Portion of existing buildings with baseline technology already
installed
- Portion baseline installed up for replacement
* Portion baseline replacements adopting another baseline technology
(#3)
* Portion baseline replacements changing to an efficient technology (#4)
- Portion baseline installed surviving (#5)

o	Portion of existing buildings with efficient technology already
installed
- Portion efficient installed up for replacement
* All efficient replacements currently assumed to adopt another
efficient technology (#6)
- Portion efficient installed surviving (#7)

For a measure that only applies to new buildings, the total
microsegment size is calculated as (1+2); the total microsegment size
for measures that only apply to existing buildings is calculated as
(3+4+5+6+7); and the total microsegment size for measures that apply to
all buildings is all of these partitions summed together.  The
‘competed’ microsegment sizes are composed of only the ‘efficient’
partitions (e.g., (2) for a measure apply to new buildings; (4+6) for a
measure that applies to existing buildings; and (2+4+6) for a measure
that applies to all buildings).

Once the total and competed microsegments are calculated as above for
the stock and energy variables, associated carbon emissions are
calculated by applying a carbon intensity factor to the energy data,
and associated costs for the stock, energy, and carbon microsegments
are finally calculated.

The ‘run_test.py’ routine is modified to accommodate the above changes.
 ‘PartitionMicrosegmentTest’ is adjusted to reflect new inputs and
outputs in the partition_microsegment function, and two new sample
measures (#9 and #10) are added to ‘FindPartitionMicrosegmentTest’ to
test cases where a measure applies to only new buildings (sample
measure #9) or only existing buildings (sample measure #10).


Former-commit-id: 9255d5894081e3206dfa41fafbb9eaa28db05798