Update annual number of total and new homes in mseg.py
For the residential sector, this commit allows the number of total and
new homes to be updated for each year in the modeling time horizon; the
home totals are broken down by census division and building type.

Leaf nodes for total and new homes numbers are now added to the
microsegments JSON hierarchy on the fuel type level, as was done
previously for square footage information.

New homes data are updated in a manner very similar to square footage
data, with the only difference being that square footage data are
included in the ‘HOUSEHOLDS’ column in RESDBOUT while new homes data
are included in the ‘EQSTOCK’ column in RESDBOUT.

Total homes data are included in the ‘HOUSEHOLDS’ column in RESDBOUT
out, but unlike square footage data, the total homes data do not have a
unique filtering identifier in the ‘ENDUSE’ column (e.g., ’SQ’ for
square footage).  Instead, total homes filtering information matches
that of electric boilers (‘HT’ for the ‘ENDUSE’ column and ‘ELEC_RAD’
for the ‘EQPCLASS’ column).  Updating of total homes data therefore
occurs in a similar fashion to electric boiler data updating, though
again total homes numbers come from the ‘HOUSEHOLDS’ column (instead of
the ‘EQSTOCK’ and ‘CONSUMPTION’ columns for electric boiler stock and
consumption data).

The mseg_test.py routine has been updated to consider new and total
homes data updating cases across all relevant tests.  A few other small
tweaks to the mseg_test.py routine were also needed to ensure that the
new ‘BULBTYPE’ column in the 2015 RESDBOUT was being handled correctly.


Former-commit-id: 3ffaabcfe5b071fc686f5611b570bc15110d4eb1