Update run.py to accept new homes numbers in mseg JSON
The microsegments JSON for the residential sector now includes
information about the number of total and new homes for each building
type and climate zone, which is available from EIA in the RESDBOUT
file.

This commit ensures that the total/new homes numbers will be correctly
discovered and used to find the fraction of energy and stock in each
microsegment that is attributable to new construction.  This fraction
is important to measures that apply only to new or existing buildings.

The ‘FindPartitionMasterMicrosegmentTest’ in run_test.py was updated to
reflect the expectation new and total homes numbers will be available
in the mseg_in input to the mseg_find_partition function.


Former-commit-id: 06cb260644f811f10710932f1d9244df04a25deb