Include measure lifetime in master microsegment dict
Scout users should be able to assign a probability distribution to a
measure’s lifetime input and/or specify different measure lifetimes for
different microsegments (e.g., one lifetime for a northern climate zone
microsegment and another for a southern climate zone microsegment).
This commit ensures that such lifetime input specification features are
correctly handled in the ‘microsegment_partition’ function of ‘run.py’.

New “baseline” and “measure” breakdowns are added to the “lifetime”
branch of the master microsegment dict.  This allows storage of the
measure lifetime information that is updated in
‘microsegment_partition’ for later use by the ’calc_metric_update’
function.  Note that the measure lifetime information that is stored in
the master microsegment dict may come in the form of a floating value
or a numpy array (if a distribution is specific on measure lifetime).

The ‘FindPartitionMasterMicrosegmentTest’ in ‘run_test.py’ has been
updated to reflect the new breakdown of lifetime information in the
master microsegment dict.  The ’calc_metric_update’ function, which
again draws from the master microsegment information, has also been
updated, as has the associated ‘PrioritizationMetricsTest’ in
‘run_test.py’.


Former-commit-id: 110d43aef7996f17de6e42cb9cc54c082ba54275