Handle array of measure output values for sample measure runs
This commit allows measure output distributions to be reported to the
CSV file.  The added code first checks whether a measure’s efficient
energy market output is an array of values rather than a point value.
If so, the average and 5th/95th percentile values for this array are
calculated and reported out as average, low, and high cases for this
output.

This process is repeated for all other outputs that are associated with
the efficient energy market output (e.g., energy savings, energy cost
savings, efficient carbon market, carbon savings, carbon cost savings).

If any of the above outputs is a point value, the average, low, and
high cases for that output are all set to the point value.


Former-commit-id: 11089f9f5b9101bff2c82c4bc7a96107249140f7