Add 'competed (captured)' stock leaf node and condense code
With this commit, a measure’s master microsegment dict now includes a
’competed (captured)’ leaf node for the ‘stock’ category.  The captured
stock value represents the number of competed stock units that are
actually captured by the measure in a given year; this value is used to
normalize measure capital, energy, and carbon costs to a per unit stock
basis in the ‘calc_metric_update’ function of ‘run.py’.

Initially, the captured portion of the competed stock is set to the
competed stock value.  In cases where measures have competition for
their baseline market, the captured stock number may be adjusted down
to reflect a less than 100% market share.  Measure market shares will
be determined in the ‘compete_measures’ function, which is to be filled
out in a future commit.

The ‘PartitionMicrosegmentTest,’ ‘FindPartitionMasterMicrosegmentTest,’
and ‘PrioritizationMetricsTest’ routines in ‘run_test.py’ have all been
updated to accommodate the new ‘competed (captured)’ stock leaf node.

In addition, a few list comprehensions have been added to ‘run.py’ to
make the code more compact.


Former-commit-id: aa743b9806a015c9ce7134e13b6331a215f37668